requirements.txt
python-dotenv==0.21.0
replicate==0.4.0
discord~=2.1.0
keytotext==2.3.2
responses~=0.18.0
requests~=2.28.1
snscrape~=0.4.3.20220106
pandas~=1.5.1